Save memory percentage as a double
git-svn-id: file:///home/svn/DVA2006@1653 1e6d2d28-e1fb-0310-b8ab-81908d8c6b6a

+++ src/java/org/deltava/commands/system/DiagnosticCommand.java
@@ -48,7 +48,7 @@ public class DiagnosticCommand extends AbstractCommand {
        ctx.setAttribute("cpuCount", new Integer(rt.availableProcessors()), REQUEST);
        ctx.setAttribute("totalMemory", new Long(rt.totalMemory()), REQUEST);
        ctx.setAttribute("maxMemory", new Long(rt.maxMemory()), REQUEST);
-       ctx.setAttribute("pctMemory", new Long(Math.round(rt.totalMemory() * 100.0 / rt.maxMemory())), REQUEST);
+       ctx.setAttribute("pctMemory", new Double(Math.round(rt.totalMemory() * 100.0 / rt.maxMemory())), REQUEST);
        
        // Get JDBC drivers
        ctx.setAttribute("jdbcDrivers", DriverManager.getDrivers(), REQUEST);
